test: PrimeFactorTestingHelperを内部メソッドに
