+++ test/matsu/num/number/primes/PrimeFactorizeTesting.java
@@ -10,6 +10,11 @@
  */
 package matsu.num.number.primes;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.*;
+import static org.hamcrest.Matchers.*;
+
+import java.util.Arrays;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.function.IntFunction;
 import java.util.function.LongFunction;
@@ -47,7 +52,7 @@ final class PrimeFactorizeTesting {
                 int n = 2 + ThreadLocalRandom.current().nextInt(Math.max(1, max()));
 
                 PrimeFactorInt primeFactor = primeFactorize.apply(n);
-                PrimeFactorTestingHelper.testInt(primeFactor);
+                testInt(primeFactor);
             }
         }
 
@@ -85,7 +90,7 @@ final class PrimeFactorizeTesting {
                 long n = 2 + ThreadLocalRandom.current().nextLong(Math.max(1L, max()));
 
                 PrimeFactorLong primeFactor = primeFactorize.apply(n);
-                PrimeFactorTestingHelper.testLong(primeFactor);
+                testLong(primeFactor);
             }
         }
 
@@ -103,6 +108,86 @@ final class PrimeFactorizeTesting {
          */
         long max() {
             return 100_000_000_000L;
+        }
+    }
+
+    /**
+     * {@link PrimeFactorInt} インスタンスが正当かどうかを検証する.
+     * 
+     * <p>
+     * 検証される項目は次の通り.
+     * </p>
+     * 
+     * <ul>
+     * <li>素因数が素数かどうか</li>
+     * <li>素因数の総積が元の数に一致するかどうか</li>
+     * <li>素因数が昇順に並んでいるかどうか</li>
+     * </ul>
+     * 
+     * @param primeFactor primeFactor
+     */
+    static void testInt(PrimeFactorInt primeFactor) {
+
+        int n = primeFactor.original();
+        int[] factors = primeFactor.factors();
+
+        int prodFactors = 1;
+
+        // 素数判定しつつ総積を計算
+        for (int f : factors) {
+            assertThat(
+                    f + "is not prime: n = " + n + ", " + Arrays.toString(factors),
+                    Primality.isPrime(f), is(true));
+            prodFactors *= f;
+        }
+
+        assertThat(
+                "product mismatch: n = " + n + ", " + Arrays.toString(factors),
+                prodFactors, is(n));
+
+        // 昇順の判定
+        for (int i = 0; i < factors.length - 1; i++) {
+            assertThat("factors is sorted?", factors[i], is(lessThanOrEqualTo(factors[i + 1])));
+        }
+    }
+
+    /**
+     * {@link PrimeFactorLong} インスタンスが正当かどうかを検証する.
+     * 
+     * <p>
+     * 検証される項目は次の通り.
+     * </p>
+     * 
+     * <ul>
+     * <li>素因数が素数かどうか</li>
+     * <li>素因数の総積が元の数に一致するかどうか</li>
+     * <li>素因数が昇順に並んでいるかどうか</li>
+     * </ul>
+     * 
+     * @param primeFactor primeFactor
+     */
+    static void testLong(PrimeFactorLong primeFactor) {
+
+        long n = primeFactor.original();
+        long[] factors = primeFactor.factors();
+
+        long prodFactors = 1;
+
+        // 素数判定しつつ総積を計算
+        for (long f : factors) {
+            assertThat(
+                    f + "is not prime: n = " + n + ", " + Arrays.toString(factors),
+                    Primality.isPrime(f), is(true));
+            prodFactors *= f;
+        }
+
+        assertThat(
+                "product mismatch: n = " + n + ", " + Arrays.toString(factors),
+                prodFactors, is(n));
+
+        // 昇順の判定
+        for (int i = 0; i < factors.length - 1; i++) {
+            assertThat("factors is sorted?", factors[i], is(lessThanOrEqualTo(factors[i + 1])));
         }
     }
 }
